chore: separate Composite and fix type errors
